requirements/test.txt
-r base.txt
Django>=1.7
djangorestframework==3.3.3
django-filter==0.13.0
